ui: Fetch the app runtime config before fetching the basename
